# src/main/resources/swagger.yml
openapi: 3.0.1
info:
  title: Gerenciamento de consultas
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /usuarios/{id}:
    get:
      tags:
        - usuario-resources
      operationId: buscarCliente_1
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Usuario"
    put:
      tags:
        - usuario-resources
      operationId: atualizarCliente_1
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Usuario"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Usuario"
    delete:
      tags:
        - usuario-resources
      operationId: deletarUsuario_1
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /consultas/{id}:
    get:
      tags:
        - consulta-resource
      operationId: buscarConsulta_1
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Consulta"
    put:
      tags:
        - consulta-resource
      operationId: atualizarConsulta_1
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Consulta"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Consulta"
    delete:
      tags:
        - consulta-resource
      operationId: deletarConsulta_1
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /usuarios:
    get:
      tags:
        - usuario-resources
      operationId: listarClientes_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Usuario"
                  - type: array
                    items:
                      $ref: "#/components/schemas/Usuario"
    post:
      tags:
        - usuario-resources
      operationId: cadastrarUsuario_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Usuario"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Usuario"
  /consultas:
    get:
      tags:
        - consulta-resource
      operationId: listarConsultar_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Consulta"
                  - type: array
                    items:
                      $ref: "#/components/schemas/Consulta"
    post:
      tags:
        - consulta-resource
      operationId: cadastrarConsulta_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Consulta"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Consulta"
components:
  schemas:
    Usuario:
      type: object
      properties:
        idUsuario:
          type: integer
          format: int64
        nomeUsuario:
          type: string
        email:
          type: string
        cpf:
          type: string
        telefone:
          type: string
        dataNascimento:
          type: string
          format: date-time
        permissao:
          type: string
          enum:
            - ADMIN
            - SECRETARIO
    Consulta:
      type: object
      properties:
        idConsulta:
          type: integer
          format: int64
        dataConsulta:
          type: string
        profissional:
          type: string
        especialidade:
          type: string
        usuario:
          $ref: "#/components/schemas/Usuario"